Fix detekt findings introduced by the defect fixes
Split the hostile iloc item count test into its own file, wrap the
GIF chunk filter chain consistently, shorten an assertion message and
use raw strings for the stale reference regexes.
